requirements.txt
numpy<=2.0.0
scipy
numba
requests
openai
torch
tensorboard
wandb
pytz
pandas
matplotlib
ttkbootstrap
gym
python-dotenv